Get rid of blivet.blivet's last deps on blivet.osinstall
The two dependencies we're eliminating here:
blivet.osinstall.StorageDiscoveryConfig
blivet.osinstall.find_existing_installations
This has the added benefit of eliminating depedence on pykickstart.constants.

Unfortunately this means changing a lot of stuff all at once, but these
are mostly moves without any changes. Stuff that's now out of
blivet.blivet.Blivet and now in blivet.osinstall.InstallerStorage:
    autopart_type (attr)
    _free_space_snapshot (attr)
    free_space_snapshot (prop)
    create_free_space_snapshot

Things which were changed in blivet.blivet.Blivet to eliminate dependence
on anything in blivet.osinstall:
    reset
    get_free_space
Those two methods exist as they originally were in
blivet.osinstall.InstallerStorage now.

InstallerStorage also handles protected devices now as well, so all of
that logic has been moved out of blivet.populator.populator and into
this class. Namely, three functions have been moved to InstallerStorage:
    _resolve_protected_device_specs
    _find_live_backing_device
    _mark_protected_device

iSCSI, FCoE, and zFCP startup/config writes now take place in
InstallerStorage.
