+++ blivet/osinstall.py
@@ -31,9 +31,9 @@ gi.require_version("BlockDev", "1.0")
 
 from gi.repository import BlockDev as blockdev
 
-from pykickstart.constants import CLEARPART_TYPE_NONE, CLEARPART_TYPE_LINUX, CLEARPART_TYPE_ALL, CLEARPART_TYPE_LIST
+from pykickstart.constants import AUTOPART_TYPE_LVM, CLEARPART_TYPE_NONE, CLEARPART_TYPE_LINUX, CLEARPART_TYPE_ALL, CLEARPART_TYPE_LIST
 
-from . import util
+from . import util, __version__
 from . import get_sysroot, get_target_physical_root, error_handler, ERROR_RAISE
 
 from .blivet import Blivet
@@ -45,9 +45,13 @@ from .errors import FSTabTypeMismatchError, UnrecognizedFSTabEntryError, Storage
 from .formats import get_device_format_class
 from .formats import get_format
 from .flags import flags
+from .iscsi import iscsi
+from .fcoe import fcoe
+from .zfcp import zfcp
 from .platform import platform as _platform
 from .platform import EFI
 from .size import Size
+from .udev import resolve_devspec
 
 from .i18n import _
 
@@ -313,7 +317,7 @@ class StorageDiscoveryConfig(object):
         self.clear_part_type = ksdata.clearpart.type
         self.clear_part_disks = ksdata.clearpart.drives[:]
         self.clear_part_devices = ksdata.clearpart.devices[:]
-        self.initialize_disks = ksdata.clearpart.init_all
+        self.initialize_disks = ksdata.clearpart.initAll
         self.zero_mbr = ksdata.zerombr.zerombr
 
 
@@ -1111,25 +1115,16 @@ class InstallerStorage(Blivet):
         self.autopart_type = AUTOPART_TYPE_LVM
 
         self.__luks_devs = {}
-
-        # these will both be empty until our reset method gets called
-        # instantiate our own devicetree here to override the default created
-        # in Blivet so that protected_dev_specs gets handled
-        self.devicetree = DeviceTree(passphrase=self.encryption_passphrase,
-                                     luks_dict=self.__luks_devs,
-                                     ignored_disks=self.ignored_disks,
-                                     exclusive_disks=self.exclusive_disks,
-                                     disk_images=self.disk_images,
-                                     protected_dev_specs=self.config.protected_dev_specs)
         self.fsset = FSSet(self.devicetree)
         self._free_space_snapshot = None
+        self.live_backing_device = None
 
     def do_it(self, callbacks=None):
         """
         Commit queued changes to disk.
 
         :param callbacks: callbacks to be invoked when actions are executed
-        :type callbacks: return value of the :func:`~.callbacks.create_new_callbacks_
+        :type callbacks: return value of the :func:`~.callbacks.create_new_callbacks_register`
 
         """
         super().do_it(callbacks=callbacks)
@@ -1217,22 +1212,18 @@ class InstallerStorage(Blivet):
             fcoe.startup()
             zfcp.startup()
 
-        self.devicetree.reset(passphrase=self.encryption_passphrase,
-                              luks_dict=self.__luks_devs,
-                              ignored_disks=self.ignored_disks,
-                              exclusive_disks=self.exclusive_disks,
-                              disk_images=self.disk_images,
-                              protected_dev_specs=self.config.protected_dev_specs)
-        self.devicetree.populate(cleanup_only=cleanup_only)
-        self.fsset = FSSet(self.devicetree)
-        self.edd_dict = get_edd_dict(self.partitioned)
-        self.devicetree.edd_dict = self.edd_dict
-        if self.bootloader:
-            # clear out bootloader attributes that refer to devices that are
-            # no longer in the tree
-            self.bootloader.reset()
+        super().reset(cleanup_only=cleanup_only)
+
+        # protected device handling
+        self.protected_dev_names = []
+        self._resolve_protected_device_specs()
+        self._find_live_backing_device()
+        for devname in self.protected_dev_names:
+            dev = self.devicetree.get_device_by_name(devname)
+            self._mark_protected_device(dev)
+
         if self.ksdata:
-             self.config.updates(self.ksdata)
+            self.config.update(self.ksdata)
         self.roots = []
         if flags.installer_mode:
             self.roots = find_existing_installations(self.devicetree)
@@ -1242,6 +1233,46 @@ class InstallerStorage(Blivet):
             self.devicetree.handle_nodev_filesystems()
 
         self.update_bootloader_disk_list()
+
+    def _resolve_protected_device_specs(self):
+        """ Resolve the protected device specs to device names. """
+        for spec in self.config.protected_dev_specs:
+            name = resolve_devspec(spec)
+            log.debug("protected device spec %s resolved to %s", spec, name)
+            if name:
+                self.protected_dev_names.append(name)
+
+    def _find_live_backing_device(self):
+        # FIXME: the backing dev for the live image can't be used as an
+        # install target.  note that this is a little bit of a hack
+        # since we're assuming that /run/initramfs/live will exist
+        for mnt in open("/proc/mounts").readlines():
+            if " /run/initramfs/live " not in mnt:
+                continue
+
+            live_device_name = mnt.split()[0].split("/")[-1]
+            log.info("%s looks to be the live device; marking as protected",
+                     live_device_name)
+            self.protected_dev_names.append(live_device_name)
+            self.live_backing_device = live_device_name
+            break
+
+    def _mark_protected_device(self, device):
+        """
+          If this device is protected, mark it as such now. Once the tree
+          has been populated, devices' protected attribute is how we will
+          identify protected devices.
+
+         :param :class: `blivet.devices.storage.StorageDevice` device: device to
+          mark as protected
+        """
+        if device.name in self.protected_dev_names:
+            device.protected = True
+            # if this is the live backing device we want to mark its parents
+            # as protected also
+            if device.name == self.live_backing_device:
+                for parent in device.parents:
+                    parent.protected = True
 
     def get_free_space(self, disks=None, clear_part_type=None):
         """ Return a dict with free space info for each disk.
@@ -1465,176 +1496,12 @@ class InstallerStorage(Blivet):
             parent = getattr(self.ksdata, list_attr)
             parent.dataList().append(data)
 
+        self.config = StorageDiscoveryConfig()
+        self.autopart_type = AUTOPART_TYPE_LVM
+
+        self.__luks_devs = {}
         self.fsset = FSSet(self.devicetree)
-
-    def reset(self, cleanup_only=False):
-        Blivet.reset(self, cleanup_only)
-
-        self.fsset = FSSet(self.devicetree)
-
-    def write(self):
-        Blivet.write(self)
-
-        self.fsset.write()
-        self.make_mtab()
-
-    @property
-    def boot_device(self):
-        dev = None
-        if self.fsset:
-            dev = self.mountpoints.get("/boot", self.root_device)
-        return dev
-
-    @property
-    def mountpoints(self):
-        return self.fsset.mountpoints
-
-    @property
-    def root_device(self):
-        return self.fsset.root_device
-
-    @property
-    def file_system_free_space(self):
-        """ Combined free space in / and /usr as :class:`~.size.Size`. """
-        mountpoints = ["/", "/usr"]
-        free = Size(0)
-        btrfs_volumes = []
-        for mountpoint in mountpoints:
-            device = self.mountpoints.get(mountpoint)
-            if not device:
-                continue
-
-            # don't count the size of btrfs volumes repeatedly when multiple
-            # subvolumes are present
-            if isinstance(device, BTRFSSubVolumeDevice):
-                if device.volume in btrfs_volumes:
-                    continue
-                else:
-                    btrfs_volumes.append(device.volume)
-
-            if device.format.exists:
-                free += device.format.free
-            else:
-                free += device.format.free_space_estimate(device.size)
-
-        return free
-    def update_ksdata(self):
-        """ Update ksdata to reflect the settings of this Blivet instance. """
-        if not self.ksdata or not self.mountpoints:
-            return
-
-        # clear out whatever was there before
-        self.ksdata.partition.partitions = []
-        self.ksdata.logvol.lvList = []
-        self.ksdata.raid.raidList = []
-        self.ksdata.volgroup.vgList = []
-        self.ksdata.btrfs.btrfsList = []
-
-        # iscsi?
-        # fcoe?
-        # zfcp?
-        # dmraid?
-
-        # bootloader
-
-        # ignoredisk
-        if self.ignored_disks:
-            self.ksdata.ignoredisk.drives = self.ignored_disks[:]
-        elif self.exclusive_disks:
-            self.ksdata.ignoredisk.onlyuse = self.exclusive_disks[:]
-
-        # autopart
-        self.ksdata.autopart.autopart = self.do_autopart
-        self.ksdata.autopart.type = self.autopart_type
-        self.ksdata.autopart.encrypted = self.encrypted_autopart
-
-        # clearpart
-        self.ksdata.clearpart.type = self.config.clear_part_type
-        self.ksdata.clearpart.drives = self.config.clear_part_disks[:]
-        self.ksdata.clearpart.devices = self.config.clear_part_devices[:]
-        self.ksdata.clearpart.initAll = self.config.initialize_disks
-        if self.ksdata.clearpart.type == CLEARPART_TYPE_NONE:
-            # Make a list of initialized disks and of removed partitions. If any
-            # partitions were removed from disks that were not completely
-            # cleared we'll have to use CLEARPART_TYPE_LIST and provide a list
-            # of all removed partitions. If no partitions were removed from a
-            # disk that was not cleared/reinitialized we can use
-            # CLEARPART_TYPE_ALL.
-            self.ksdata.clearpart.devices = []
-            self.ksdata.clearpart.drives = []
-            fresh_disks = [d.name for d in self.disks if d.partitioned and
-                           not d.format.exists]
-
-            destroy_actions = self.devicetree.actions.find(action_type="destroy",
-                                                           object_type="device")
-
-            cleared_partitions = []
-            partial = False
-            for action in destroy_actions:
-                if action.device.type == "partition":
-                    if action.device.disk.name not in fresh_disks:
-                        partial = True
-
-                    cleared_partitions.append(action.device.name)
-
-            if not destroy_actions:
-                pass
-            elif partial:
-                # make a list of removed partitions
-                self.ksdata.clearpart.type = CLEARPART_TYPE_LIST
-                self.ksdata.clearpart.devices = cleared_partitions
-            else:
-                # if they didn't partially clear any disks, use the shorthand
-                self.ksdata.clearpart.type = CLEARPART_TYPE_ALL
-                self.ksdata.clearpart.drives = fresh_disks
-
-        if self.do_autopart:
-            return
-
-        self._update_custom_storage_ksdata()
-
-    def _update_custom_storage_ksdata(self):
-        """ Update KSData for custom storage. """
-
-        # custom storage
-        ks_map = {PartitionDevice: ("PartData", "partition"),
-                  TmpFSDevice: ("PartData", "partition"),
-                  LVMLogicalVolumeDevice: ("LogVolData", "logvol"),
-                  LVMVolumeGroupDevice: ("VolGroupData", "volgroup"),
-                  MDRaidArrayDevice: ("RaidData", "raid"),
-                  BTRFSDevice: ("BTRFSData", "btrfs")}
-
-        # make a list of ancestors of all used devices
-        devices = list(set(a for d in list(self.mountpoints.values()) + self.swaps
-                           for a in d.ancestors))
-
-        # devices which share information with their distinct raw device
-        complementary_devices = [d for d in devices if d.raw_device is not d]
-
-        devices.sort(key=lambda d: len(d.ancestors))
-        for device in devices:
-            cls = next((c for c in ks_map if isinstance(device, c)), None)
-            if cls is None:
-                log.info("omitting ksdata: %s", device)
-                continue
-
-            class_attr, list_attr = ks_map[cls]
-
-            cls = getattr(self.ksdata, class_attr)
-            data = cls()    # all defaults
-
-            complements = [d for d in complementary_devices if d.raw_device is device]
-
-            if len(complements) > 1:
-                log.warning("omitting ksdata for %s, found too many (%d) complementary devices", device, len(complements))
-                continue
-
-            device = complements[0] if complements else device
-
-            device.populate_ksdata(data)
-
-            parent = getattr(self.ksdata, list_attr)
-            parent.dataList().append(data)
+        self._free_space_snapshot = None
 
     def empty_device(self, device):
         empty = True
@@ -1874,6 +1741,18 @@ class InstallerStorage(Blivet):
             os.unlink(path)
 
         os.symlink(target, path)
+
+    @property
+    def free_space_snapshot(self):
+        # if no snapshot is available, do it now and return it
+        self._free_space_snapshot = self._free_space_snapshot or self.get_free_space()
+
+        return self._free_space_snapshot
+
+    def create_free_space_snapshot(self):
+        self._free_space_snapshot = self.get_free_space()
+
+        return self._free_space_snapshot
 
     def add_fstab_swap(self, device):
         """
